+++ src/main/java/com/ozgurbaybas/Repository/JobTitleRepository.java
@@ -3,6 +3,6 @@ package com.ozgurbaybas.Repository;
 import com.ozgurbaybas.Models.JobTitle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobTitleRepository extends JpaRepository<JobTitle, Long> {
+public interface JobTitleRepository extends JpaRepository<JobTitle, Integer> {
     JobTitle getByTitle(String title);
 }
